+++ src/Light/Geometry/Vector.hs
@@ -16,14 +16,14 @@ where
 
 import Control.Lens
 
-data Vector = Vector { _dx :: Float, _dy :: Float, _dz :: Float } deriving (Show, Read)
+data Vector = Vector { _dx :: Double, _dy :: Double, _dz :: Double } deriving (Show, Read)
 
-vector :: Float -> Float -> Float -> Vector
+vector :: Double -> Double -> Double -> Vector
 vector = Vector
 
 makeLenses ''Vector
 
-ds :: Lens' Vector [Float]
+ds :: Lens' Vector [Double]
 ds = lens (\ (Vector x y z) -> [x, y, z, 0]) (\_ [x, y, z, 0] -> Vector x y z)
 
 instance Eq Vector where
@@ -39,14 +39,14 @@ unitZVector = Vector 0 0 1
 (Vector x y z) ^+^ (Vector a b c) = Vector (x+a) (y+b) (z+c)
 (Vector x y z) ^-^ (Vector a b c) = Vector (x-a) (y-b) (z-c)
 
-(^*), (^/) :: Vector -> Float -> Vector
+(^*), (^/) :: Vector -> Double -> Vector
 (Vector x y z) ^* s = Vector (x*s) (y*s) (z*s)
 (Vector x y z) ^/ s = Vector (x/s) (y/s) (z/s)
 
-(*^) :: Float -> Vector -> Vector
+(*^) :: Double -> Vector -> Vector
 s *^ (Vector x y z) = Vector (x*s) (y*s) (z*s)
 
-(^.^) :: Vector -> Vector -> Float
+(^.^) :: Vector -> Vector -> Double
 (Vector x y z) ^.^ (Vector a b c) = (x*a) + (y*b) + (z*c)
 
 infixl 6 ^+^, ^-^
@@ -55,7 +55,7 @@ infixl 7 ^.^, ^*, ^/, *^
 negateV :: Vector -> Vector
 negateV (Vector x y z) = Vector (-x) (-y) (-z)
 
-magnitudeV, magnitudeSquaredV :: Vector -> Float
+magnitudeV, magnitudeSquaredV :: Vector -> Double
 magnitudeV = sqrt . magnitudeSquaredV
 magnitudeSquaredV v = v ^.^ v
 
@@ -70,5 +70,5 @@ cross (Vector ux uy uz) (Vector vx vy vz) =
          (uz*vx - ux*vz)
          (ux*vy - uy*vx)
 
-angleBetween :: Vector -> Vector -> Float
+angleBetween :: Vector -> Vector -> Double
 angleBetween v w = acos $ (v ^.^ w) / (magnitudeV v * magnitudeV w)
